requirements.txt
streamlit==1.32.0
tflite-runtime==2.12.0
numpy==1.24.4
Pillow==9.5.0
>>>>>>> ad228a3 (Fix deployment error: replace TensorFlow with tflite-runtime for lighter dependencies)
